+++ app/test/package/package_publisher_test.dart
@@ -295,7 +295,7 @@ void main() {
   });
 }
 
-dynamic _json(value) => json.decode(json.encode(value));
+dynamic _json(Object? value) => json.decode(json.encode(value));
 
 void _testUserNotMemberOfPublisher({
   required Future<void> Function(PubApiClient client) fn,