+++ app/test/publisher/api_test.dart
@@ -989,7 +989,7 @@ void main() {
   });
 }
 
-dynamic _json(value) => json.decode(json.encode(value));
+dynamic _json(Object? value) => json.decode(json.encode(value));
 
 void _testAdminAuthIssues(Future Function(PubApiClient client) fn) {
   setupTestsWithCallerAuthorizationIssues(fn);