+++ CHANGELOG.md
@@ -4,6 +4,8 @@ AppEngine version, listed here to ease deployment and troubleshooting.
 ## Next Release (replace with git tag when deployed)
  * Bump runtimeVersion to `2025.09.08`.
  * Upgraded stable Flutter analysis SDK to `3.35.3`.
+ * Upgraded pana to `0.22.23`.
+ * Upgrade `package:analyzer` to `^8.0.0`.
  * Updated SDK constraint (and language version) to `3.9`
 
 ## `20250904t074800-all`